Move acceptance test utility classes to common file
