+++ tests/atest/transformers/AddMissingEnd/test_transformer.py
@@ -1,4 +1,4 @@
-from .. import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class TestAddMissingEnd(TransformerAcceptanceTest):
