+++ tests/atest/transformers/AlignTestCasesSection/test_transformer.py
@@ -2,7 +2,7 @@ from typing import Tuple
 
 import pytest
 
-from .. import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class TestAlignTestCasesSection(TransformerAcceptanceTest):
