+++ tests/atest/transformers/MergeAndOrderSections/test_merge_and_order_sections.py
@@ -1,6 +1,6 @@
 import pytest
 
-from .. import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class TestMergeAndOrderSections(TransformerAcceptanceTest):
