+++ tests/atest/transformers/NormalizeNewLines/test_normalize_new_lines.py
@@ -1,6 +1,6 @@
 import pytest
 
-from .. import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class TestNormalizeNewLines(TransformerAcceptanceTest):
