+++ tests/atest/transformers/NormalizeSectionHeaderName/test_normalize_section_header.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.atest.transformers import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class TestNormalizeSectionHeaderName(TransformerAcceptanceTest):
