+++ tests/atest/transformers/NormalizeSettingName/test_normalize_setting_name.py
@@ -1,4 +1,4 @@
-from .. import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class TestNormalizeSettingName(TransformerAcceptanceTest):
