+++ tests/atest/transformers/OrderSettings/test_order_settings.py
@@ -1,6 +1,6 @@
 import pytest
 
-from .. import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class TestOrderSettings(TransformerAcceptanceTest):
