+++ tests/atest/transformers/ReplaceBreakContinue/test_transformer.py
@@ -1,4 +1,4 @@
-from .. import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class TestReplaceBreakContinue(TransformerAcceptanceTest):
