+++ tests/atest/transformers/ReplaceReturns/test_transformer.py
@@ -1,6 +1,6 @@
 import pytest
 
-from .. import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class TestReplaceReturns(TransformerAcceptanceTest):
