+++ utils/new_transformer_templates/test_transformer.template
@@ -1,4 +1,4 @@
-from .. import TransformerAcceptanceTest
+from tests.atest import TransformerAcceptanceTest
 
 
 class Test{{ transformer_name }}(TransformerAcceptanceTest):
